fix: scroll is faster

+++ Kernel/include/console.h
@@ -6,5 +6,5 @@ void putChar(char c);
 uint64_t print(char* string, uint64_t count);
 uint64_t readLine(char* buffer, uint64_t max);
 void puts(char* string);
-void set_text_size(float size);
+void set_text_size(uint64_t size);
 void redraw();
